* behoben: doppelte Grußformel aus Mail-Template entfernt * hinzugefügt: Migrations für error logging * hinzugefügt: routing branch für interne Funktionalität * hinzugefügt: backend error logging im allgemeinen error handler

+++ db/app_migrations/034_create_frontend_errors.rb
@@ -9,12 +9,16 @@ class CreateFrontendErrors < Sequel::Migration
         String :backend_version
         String :message
         String :stacktrace
+        String :ip_hash
 
         DateTime :created_at
       end
     end
   
+  
     def down
-      drop_table :frontend_errors
+      if @db.table_exists?(:frontend_errors)
+        drop_table :frontend_errors
+      end
     end
   end
